Remove event rss ingestor
