+++ lib/ingestors/ingestor_factory.rb
@@ -13,7 +13,6 @@ module Ingestors
         Ingestors::ZenodoIngestor,
         Ingestors::OaiPmhIngestor,
         Ingestors::GithubIngestor,
-        Ingestors::EventRSSIngestor,
         Ingestors::MaterialRSSIngestor
       ] + taxila_ingestors + llm_ingestors + heptraining_ingestors
     end
