+++ lib/ingestors/youtube_ingestor.rb
@@ -11,22 +11,19 @@ module Ingestors
     end
 
     def discover_feed_url(content, base_url)
-      super(content, base_url)
+      super_url = super(content, base_url) # discovers url from HTML
+      return super_url if super_url
 
-      return unless (url = discover_feed_url_from_youtube_playlist_url(base_url))
-
-      @messages << "Found Atom feed link from YouTube playlist URL, following: #{url}"
-      url
-    end
-
-    def discover_feed_url_from_youtube_playlist_url(base_url)
+      # YouTube does not include feed URL of playlists in HTML
       uri = URI.parse(base_url)
       return nil unless Renderers::Youtube.is_youtube_url?(base_url)
 
       playlist_id = CGI.parse(uri.query.to_s).fetch('list', []).first
       return nil if playlist_id.blank?
 
-      "https://www.youtube.com/feeds/videos.xml?playlist_id=#{CGI.escape(playlist_id)}"
+      url = "https://www.youtube.com/feeds/videos.xml?playlist_id=#{CGI.escape(playlist_id)}"
+      @messages << "Found Atom feed link from YouTube playlist URL, following: #{url}"
+      url
     rescue URI::InvalidURIError
       nil
     end
